+++ src/features/navigation/NavBar.tsx
@@ -247,6 +247,29 @@ export function NavBar() {
                     </Link>
                   )}
                 </Menu.Item>
+
+                <Menu.Item>
+                  {({ active }) => (
+                    <Link
+                      href={NAV_ITEMS.ampVsRes.href}
+                      className={cn(
+                        active ? 'bg-gray-800' : '',
+                        'flex w-full flex-row items-start justify-start p-2 text-sm font-semibold text-white',
+                      )}
+                    >
+                      <div className="mr-4 w-[20px]">
+                        <NAV_ITEMS.ampVsRes.icon className="h-5 w-5 text-green-600" />
+                      </div>
+                      <div className="flex flex-col items-start justify-start gap-y-1">
+                        {NAV_ITEMS.ampVsRes.label}
+                        <p className="text-xs font-normal text-gray-400">
+                          {NAV_ITEMS.ampVsRes.description}
+                        </p>
+                      </div>
+                    </Link>
+                  )}
+                </Menu.Item>
+
                 <Menu.Item>
                   {({ active }) => (
                     <Link
@@ -269,6 +292,7 @@ export function NavBar() {
                     </Link>
                   )}
                 </Menu.Item>
+
                 <Menu.Item>
                   {({ active }) => (
                     <Link
@@ -425,6 +449,8 @@ export function NavBar() {
                   </div>
                 </Link>
 
+                <hr className="border-purple-900" />
+
                 <Link
                   href={NAV_ITEMS.bossTracker.href}
                   className="flex flex-row items-center justify-start"
@@ -459,6 +485,8 @@ export function NavBar() {
                   </div>
                 </Link>
 
+                <hr className="border-purple-900" />
+
                 <Link
                   href={NAV_ITEMS.itemLookup.href}
                   className="flex flex-row items-center justify-start"
@@ -472,6 +500,23 @@ export function NavBar() {
 
                     <p className="text-xs text-gray-400">
                       {NAV_ITEMS.itemLookup.description}
+                    </p>
+                  </div>
+                </Link>
+
+                <Link
+                  href={NAV_ITEMS.ampVsRes.href}
+                  className="flex flex-row items-center justify-start"
+                >
+                  <NAV_ITEMS.ampVsRes.icon
+                    className="mr-2 h-7 w-5 flex-none text-green-600"
+                    aria-hidden="true"
+                  />
+                  <div className="flex flex-col items-start justify-start px-3 py-2">
+                    {NAV_ITEMS.ampVsRes.label}
+
+                    <p className="text-xs text-gray-400">
+                      {NAV_ITEMS.ampVsRes.description}
                     </p>
                   </div>
                 </Link>
@@ -515,6 +560,8 @@ export function NavBar() {
                     </p>
                   </div>
                 </Link>
+
+                <hr className="border-purple-900" />
 
                 <Link
                   href={NAV_ITEMS.supportR2TK.href}