+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-02-23
+
+### Added
+
+- Added link to Amplitude vs Resonance guide under Resources.
+
 ## 2024-02-22
 
 ### Added